ThemeDetailPage 리팩토링
